Add notion of list-like children with LSLASTNodeList
One of the problems with all of the nodes being intrusively linked
is that if you want to have list-like children, all of their
siblings will be added as direct children of the parent node.

That messes with nodes like the for statement where you have
the first child being a list of expressions, an expression, then
a last expression list child. Without a node to group all of
the initializer expressions together, we don't have a predictable
child index we can use to address the condition expression or the
increment expression list.

This also fixes the issue where global storages and state were
all mashed together into one big child list under an LSLScript node.

Doesn't currently touch cases where something that would logically
be a variable-sized list of nodes is the last child of a node,
like for function call expressions.